Refactor: Rename `soroscope-api` crate to `api` and implement graceful shutdown and health check with updated dependencies.
