+++ api/Cargo.toml
@@ -1,11 +1,13 @@
 [package]
-name = "soroscope-api"
+name = "api"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-axum = "0.7"
-tokio = { version = "1", features = ["full"] }
+axum = "0.7.5"
+tokio = { version = "1.37.0", features = ["full"] }
+serde = { version = "1.0.197", features = ["derive"] }
+serde_json = "1.0.115"
 tower = "0.4"
 tower-http = { version = "0.5", features = ["cors"] }
 tracing = "0.1"
