+++ api/src/main.rs
@@ -1,13 +1,25 @@
-use axum::{routing::get, Router};
+use axum::{
+    routing::get,
+    Json, Router,
+};
+use serde_json::{json, Value};
+use std::net::SocketAddr;
+use tokio::signal;
 use tower_http::cors::CorsLayer;
 
 #[tokio::main]
 async fn main() {
+    // Initialize tracing
     tracing_subscriber::fmt::init();
 
+    // Port configuration
+    let port = std::env::var("PORT")
+        .ok()
+        .and_then(|p| p.parse().ok())
+        .unwrap_or(8080);
+    let addr = SocketAddr::from(([0, 0, 0, 0], port));
+
     // Configure CORS
-    // Allow localhost:3000 for development
-    // TODO: Add production domain when available
     let cors = CorsLayer::new()
         .allow_methods([axum::http::Method::GET, axum::http::Method::POST])
         .allow_headers([axum::http::header::CONTENT_TYPE])
@@ -17,12 +29,46 @@ async fn main() {
                 .unwrap(),
         );
 
+    // Build our application with a route
     let app = Router::new()
         .route("/", get(|| async { "SoroScope API" }))
-        .route("/health", get(|| async { "OK" }))
+        .route("/health", get(health_check))
         .layer(cors);
 
-    let listener = tokio::net::TcpListener::bind("0.0.0.0:8080").await.unwrap();
-    tracing::info!("Server running on http://0.0.0.0:8080");
-    axum::serve(listener, app).await.unwrap();
+    tracing::info!("listening on {}", addr);
+    let listener = tokio::net::TcpListener::bind(&addr).await.unwrap();
+    axum::serve(listener, app)
+        .with_graceful_shutdown(shutdown_signal())
+        .await
+        .unwrap();
+}
+
+async fn health_check() -> Json<Value> {
+    Json(json!({ "status": "ok" }))
+}
+
+async fn shutdown_signal() {
+    let ctrl_c = async {
+        signal::ctrl_c()
+            .await
+            .expect("failed to install Ctrl+C handler");
+    };
+
+    #[cfg(unix)]
+    let terminate = async {
+        signal::unix::signal(signal::unix::SignalKind::terminate())
+            .expect("failed to install signal handler")
+            .recv()
+            .await;
+    };
+
+    #[cfg(not(unix))]
+    let terminate = std::future::pending::<()>();
+
+    tokio::select! {
+        _ = ctrl_c => {},
+        _ = terminate => {},
+    }
+
+    println!("signal received, starting graceful shutdown");
 }
