--- soroscope-api/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "soroscope-api"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-axum = "0.7.5"
-tokio = { version = "1.37.0", features = ["full"] }
-serde = { version = "1.0.197", features = ["derive"] }
-serde_json = "1.0.115"
