--- soroscope-api/src/main.rs
@@ -1,58 +1,0 @@
-use axum::{
-    routing::get,
-    Json, Router,
-};
-use serde_json::{json, Value};
-use std::net::SocketAddr;
-use tokio::signal;
-
-#[tokio::main]
-async fn main() {
-    // Port configuration
-    let port = std::env::var("PORT")
-        .ok()
-        .and_then(|p| p.parse().ok())
-        .unwrap_or(8080);
-    let addr = SocketAddr::from(([0, 0, 0, 0], port));
-
-    // Build our application with a route
-    let app = Router::new()
-        .route("/health", get(health_check));
-
-    println!("listening on {}", addr);
-    let listener = tokio::net::TcpListener::bind(&addr).await.unwrap();
-    axum::serve(listener, app)
-        .with_graceful_shutdown(shutdown_signal())
-        .await
-        .unwrap();
-}
-
-async fn health_check() -> Json<Value> {
-    Json(json!({ "status": "ok" }))
-}
-
-async fn shutdown_signal() {
-    let ctrl_c = async {
-        signal::ctrl_c()
-            .await
-            .expect("failed to install Ctrl+C handler");
-    };
-
-    #[cfg(unix)]
-    let terminate = async {
-        signal::unix::signal(signal::unix::SignalKind::terminate())
-            .expect("failed to install signal handler")
-            .recv()
-            .await;
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        _ = ctrl_c => {},
-        _ = terminate => {},
-    }
-
-    println!("signal received, starting graceful shutdown");
-}
